PUMA-135 tests for dblock; some code organisation
