+++ experiments/consts.py
@@ -1,0 +1,12 @@
+# coding=utf-8
+
+
+CONTROL_STATE = 0
+ENABLED_STATE = 1
+TRACK_STATE = 3
+
+STATES = (
+    (CONTROL_STATE, 'Default/Control'),
+    (ENABLED_STATE, 'Enabled'),
+    (TRACK_STATE, 'Track'),
+)
