--- experiments/reset_intermediary.py
@@ -1,4 +1,0 @@
-from experiments.counters import Counters
-c = Counters()
-c.reset_pattern('intermediary:*')
-
